Migrate ListView to MVVM
